backend/pyproject.toml
[project]
name = "doc-gen-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "boto3>=1.42.6",
    "boto3-stubs>=1.42.7",
    "contextgem>=0.19.4",
    "fastapi[standard]>=0.119.0",
    "openai>=2.7.2",
    "openpyxl>=3.1.5",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.14.4",
    "sqlalchemy>=2.0.44",
]

[tool.ruff]
indent-width = 4
exclude = ["prompts/*.py"]


[tool.ruff.lint]
# Add the `line-too-long` rule to the enforced rule set. By default, Ruff omits rules that
# overlap with the use of a formatter, like Black, but we can override this behavior by
# explicitly adding the rule.

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
